move Resolver specific types into the Resolve module
