+++ test/resolve.jl
@@ -7,7 +7,7 @@ using Test
 using Pkg.Types
 using Pkg.Types: VersionBound
 using Pkg.Resolve
-import Pkg.Resolve: VersionWeight, add_reqs!, simplify_graph!
+import Pkg.Resolve: VersionWeight, add_reqs!, simplify_graph!, ResolverError, Fixed, Requires
 import Pkg.Types: uuid5, uuid_package
 
 # print info, stats etc.
